fix: code review address and move test cases in single file
